requirements.txt
autopep8==1.4.3
backports.csv==1.0.7
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-import-export==1.2.0
django-jet==1.0.8
django-rest-framework==0.1.0
djangorestframework==3.9.2
djangorestframework-csv==2.1.0
djangorestframework-stubs==0.4.0
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.1
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pytz==2018.9
PyYAML==5.1
six==1.12.0
tablib==0.13.0
unicodecsv==0.14.1
xlrd==1.2.0
xlwt==1.3.0
